backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
motor==3.6.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
stripe==10.0.0
python-multipart==0.0.12
slowapi==0.1.9
redis==5.2.0
minio==7.2.0
